iso3166-check-for-updates/requirements.txt
iso3166-updates
iso3166
google-cloud-storage
requests
emoji-country-flag